Move Result to Razer namespace.
